+++ relay-dynamic-config/src/defaults.rs
@@ -41,7 +41,10 @@ const RESOURCE_SPAN_OPS: &[&str] = &["resource.script", "resource.css", "resourc
 /// This requires the `SpanMetricsExtraction` feature. This feature should be set to `false` if the
 /// default should no longer be placed.
 pub fn add_span_metrics(project_config: &mut ProjectConfig) {
-    if !project_config.features.has(Feature::SpanMetricsExtraction) {
+    if !project_config
+        .features
+        .has(Feature::ExtractSpansAndSpanMetricsFromEvent)
+    {
         return;
     }
 
@@ -131,7 +134,7 @@ fn span_metrics() -> impl IntoIterator<Item = MetricSpec> {
             category: DataCategory::Span,
             mri: "c:spans/usage@none".into(),
             field: None,
-            condition: Some(exclusive_time_light_condition.clone()),
+            condition: None,
             tags: vec![],
         },
         MetricSpec {