+++ relay-server/src/statsd.rs
@@ -20,6 +20,14 @@ pub enum RelayGauges {
     ///
     /// The disk buffer size can be configured with `spool.envelopes.max_disk_size`.
     BufferEnvelopesDiskCount,
+    /// The currently used memory by the entire system.
+    ///
+    /// Relay uses the same value for its memory health check.
+    SystemMemoryUsed,
+    /// The total system memory.
+    ///
+    /// Relay uses the same value for its memory health check.
+    SystemMemoryTotal,
 }
 
 impl GaugeMetric for RelayGauges {
@@ -29,6 +37,8 @@ impl GaugeMetric for RelayGauges {
             RelayGauges::ProjectCacheGarbageQueueSize => "project_cache.garbage.queue_size",
             RelayGauges::BufferEnvelopesMemoryCount => "buffer.envelopes_mem_count",
             RelayGauges::BufferEnvelopesDiskCount => "buffer.envelopes_disk_count",
+            RelayGauges::SystemMemoryUsed => "health.system_memory.used",
+            RelayGauges::SystemMemoryTotal => "health.system_memory.total",
         }
     }
 }
@@ -550,6 +560,10 @@ pub enum RelayCounters {
     ///  - `is_segment` (only for event_type span): `true` the span is the root of a segment.
     ///  - `has_parent` (only for event_type span): `false` if the span is the root of a trace.
     ///  - `platform` (only for event_type span): The platform from which the span was spent.
+    ///  - `metric_type` (only for event_type metric): The metric type, counter, distribution,
+    ///  gauge or set.
+    ///  - `metric_encoding` (only for event_type metric): The encoding used for distribution and
+    ///  set metrics.
     ///
     /// The message types can be:
     ///